Upgrade with new perftest v2 feature set
This commit adds new branch v2, adds util/flag.go and util/cw_put.go,
updates the Makefile to rebuild perftest if any source files have
changed, and improves the perftest app with the ability to publish
test results to AWS CloudWatch or via a webhook, and to send an alert if
performance exceeds some bound.

The README describes how to run the docker as on the Rafay platform.